Definition update
Artificial intelligence has brought a revolution in the industry. Companies are desperately poaching talent like anything. However, the shortage of skilled talent in the industry is a problem yet to be worked out.  This has led to sky rocketing salaries in this domain.
Let’s take a look at some figures.

Job opportunities in artificial intelligence around the world are increasing. In the past three years, the number of job opportunities have almost doubled, according to Indeed.  The search entries for ‘artificial intelligence’ and ‘machine learning’ have increased by 185 times, as per a report from Bloomberg. However, this doesn’t necessarily mean that qualified candidates are available for the job.